requirements.txt
datasets==4.6.0
pandas==3.0.1
openai==2.24.0
tiktoken==0.12.0
pyyaml==6.0.3
jsonschema==4.26.0
tqdm==4.67.3
inspect-ai==0.3.183
agentdojo==0.1.35
requests==2.32.5
zipfile36==0.1.3
pydantic==2.12.5
python-dotenv==1.2.1